add schemas/types for argo event types created from openapi spec + to remove circular dependency
