+++ src/schemas/index.ts
@@ -1,2 +1,3 @@
 export * from './common';
 export * from './sqs';
+export * from './types';
